Feat: 역할 제거 요청 시 자원이 없다면 예외 처리

+++ server/src/main/java/com/ryc/api/v2/role/service/RoleService.java
@@ -63,6 +63,10 @@ public class RoleService {
   @Transactional
   @HasRole(Role.OWNER)
   public void deleteRole(ClubRoleSecuredDto dto, String targetUserId) {
+    if (!roleRepository.existsByAdminIdAndClubId(dto.clubId(), targetUserId)) {
+      throw new ClubException(ClubErrorCode.CLUB_MEMBER_NOT_FOUND);
+    }
+
     roleRepository.deleteByUserId(targetUserId);
   }
 }
